requirements.txt
opencv-python
pytesseract
Pillow
piexif
